+++ R/xyzm.R
@@ -68,7 +68,8 @@ as_xy <- function(x, ...) {
 #' @rdname xy
 #' @export
 as_xy.default <- function(x, ..., dims = NULL) {
-  result <- wk_translate(x, new_wk_xy(crs = wk_crs_inherit()))
+  result <- wk_handle(x, xy_writer())
+  wk_crs(result) <- wk_crs(x)
 
   if (is.null(dims)) {
     result
@@ -173,6 +174,12 @@ as_xy.matrix <- function(x, ..., crs = NULL) {
 #' @rdname xy
 #' @export
 as_xy.data.frame <- function(x, ..., dims = NULL, crs = NULL) {
+  col_handleable <- vapply(x, is_handleable, logical(1))
+  if (any(col_handleable)) {
+    stopifnot(missing(crs))
+    return(as_xy.default(x[[which(col_handleable)[1]]], dims = dims))
+  }
+
   if (is.null(dims)) {
     dims <- intersect(c("x", "y", "z", "m"), names(x))
   }